[swing] Move swing interop related functionality in separate package
